+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.2
+
+- Added styles generator with support for SASS and Stylus
+
+## 0.5.0
+
+- Added cli and layout generators
+
 ## 0.4.0
 
 - All tests pass :)